nvim: Scope LSP actions to '<leader>l' to free up `g` set of commands
This is conflicting with the useful `gT` or `gT` and in general, there are way
too many useful things around `g` to avoid using.

The downside is that there's one more key to press (`<leader>lt` instead of
`gt`) for some of the LSP actions but this seems like a fair trade off.