nvim: Disable relativenumber
The idea of jumping n lines up or down seemed useful but it does't work for me
as these jumps are "fragile". If I move a bit than I'm already off, and with
page jumps it is even harder.

It takes less effort for me to jump to a line number using `155G` than
remembering the relative number to jump.